tests: update in test coverage
