--- projects/melon-runtime/Melon.Tests/Melon/data/DataTests.cs
@@ -1,36 +1,0 @@
-﻿using Jint;
-using Melon.Engine.Builders;
-using Xunit;
-
-namespace Melon.Tests.Melon.std
-{
-    public class DataTests
-    {
-        private readonly Jint.Engine _engine;
-
-        public DataTests()
-        {
-            var builder = new EngineBuilder();
-            builder.Load("Bundle/core");
-            _engine = builder.Build();
-        }
-
-        [Fact(DisplayName = "Melon.std.boolean.checkOne should work correctly")]
-        public void StdBooleanCheckOneShouldWorkCorrectly()
-        {
-            var script = @"Melon.std.boolean.checkOne(x => x === true, [true, false, false])";
-            var result = _engine.Evaluate(script).AsBoolean();
-
-            Assert.True(result);
-        }
-
-        [Fact(DisplayName = "Melon.std.boolean.checkAll should work correctly")]
-        public void StdBooleanCheckAllShouldWorkCorrectly()
-        {
-            var script = @"Melon.std.boolean.checkAll(x => x === true, [true, true, true])";
-            var result = _engine.Evaluate(script).AsBoolean();
-
-            Assert.True(result);
-        }
-    }
-}
